b/317922302 Split Asset Inventory clients (#233)
* Split client into PolicyAnalyzerClient, AssetInventoryClient
* Use Singleton injection to allow subclasses
