+++ sources/src/main/java/com/google/solutions/jitaccess/core/catalog/TokenSigner.java
@@ -27,6 +27,7 @@ import com.google.solutions.jitaccess.core.AccessException;
 import com.google.solutions.jitaccess.core.UserId;
 import com.google.solutions.jitaccess.core.clients.IamCredentialsClient;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Singleton;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -35,7 +36,7 @@ import java.time.Instant;
 /**
  * Signs JWTs using a service account's Google-managed service account key.
  */
-@ApplicationScoped
+@Singleton
 public class TokenSigner {
   private final IamCredentialsClient iamCredentialsClient;
   private final Options options;
